+++ src/test/java/laddergame/domain/ladder/LadderHeightTest.java
@@ -1,4 +1,4 @@
-package laddergame.domain.height;
+package laddergame.domain.ladder;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
